I believe php is mostly working, still some issue with REST API in wordpress but idk man, fixed vulnerability that made it possible to run as many hanging php sessions as a user wanted

+++ http_header.h
@@ -44,7 +44,7 @@ typedef struct {
     connection_type connection;
     method_type method;
 
-    unsigned int content_length;
+    long content_length;
 
     char* resource;
     char* version;
